feat(leebrary): organize files
